src/requirements.txt
pyspark~=3.5.3
PyGithub~=2.5.0
pyYAML~=6.0.2